requirements.txt
python-dotenv==1.0.0
python_telegram_bot==20.2
telegram==0.0.1
